+++ nflow-engine/src/main/java/io/nflow/engine/workflow/definition/WorkflowSettings.java
@@ -2,11 +2,12 @@ package io.nflow.engine.workflow.definition;
 
 import static java.lang.Math.max;
 import static java.lang.Math.min;
-import static java.util.concurrent.TimeUnit.DAYS;
-import static java.util.concurrent.TimeUnit.MINUTES;
-import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.joda.time.DateTime.now;
 import static org.joda.time.DateTimeUtils.currentTimeMillis;
+import static org.joda.time.Duration.millis;
+import static org.joda.time.Duration.standardDays;
+import static org.joda.time.Duration.standardMinutes;
+import static org.joda.time.Duration.standardSeconds;
 import static org.slf4j.LoggerFactory.getLogger;
 
 import java.math.BigInteger;
@@ -18,7 +19,10 @@ import java.util.function.BiFunction;
 import java.util.function.BooleanSupplier;
 
 import org.joda.time.DateTime;
+import org.joda.time.Duration;
 import org.joda.time.LocalDateTime;
+import org.joda.time.ReadableDuration;
+import org.joda.time.Period;
 import org.joda.time.ReadablePeriod;
 import org.slf4j.Logger;
 
@@ -34,19 +38,15 @@ public class WorkflowSettings extends ModelObject {
   /**
    * Minimum delay on execution retry after an error. Unit is milliseconds.
    */
-  public final int minErrorTransitionDelay;
+  public final long minErrorTransitionDelay;
   /**
    * Maximum delay on execution retry after an error. Unit is milliseconds.
    */
-  public final int maxErrorTransitionDelay;
+  public final long maxErrorTransitionDelay;
   /**
    * Length of forced delay to break execution of a step that is considered to be busy looping. Unit is milliseconds.
    */
-  public final int shortTransitionDelay;
-  /**
-   * Immediate transition delay.
-   */
-  public final int immediateTransitionDelay;
+  public final long shortTransitionDelay;
   /**
    * Maximum retry attempts.
    */
@@ -82,10 +82,9 @@ public class WorkflowSettings extends ModelObject {
   private static final Logger logger = getLogger(WorkflowSettings.class);
 
   WorkflowSettings(Builder builder) {
-    this.minErrorTransitionDelay = builder.minErrorTransitionDelay;
-    this.maxErrorTransitionDelay = builder.maxErrorTransitionDelay;
-    this.shortTransitionDelay = builder.shortTransitionDelay;
-    this.immediateTransitionDelay = builder.immediateTransitionDelay;
+    this.minErrorTransitionDelay = builder.minErrorTransitionDelay.getMillis();
+    this.maxErrorTransitionDelay = builder.maxErrorTransitionDelay.getMillis();
+    this.shortTransitionDelay = builder.shortTransitionDelay.getMillis();
     this.maxRetries = builder.maxRetries;
     this.maxSubsequentStateExecutions = builder.maxSubsequentStateExecutions;
     this.maxSubsequentStateExecutionsPerState = new HashMap<>(builder.maxSubsequentStateExecutionsPerState);
@@ -100,14 +99,13 @@ public class WorkflowSettings extends ModelObject {
    */
   public static class Builder {
 
-    int maxErrorTransitionDelay = (int) DAYS.toMillis(1);
-    int minErrorTransitionDelay = (int) MINUTES.toMillis(1);
-    int shortTransitionDelay = (int) SECONDS.toMillis(30);
-    int immediateTransitionDelay = 0;
+    ReadableDuration maxErrorTransitionDelay = standardDays(1);
+    ReadableDuration minErrorTransitionDelay = standardMinutes(1);
+    ReadableDuration shortTransitionDelay = standardSeconds(30);
     int maxRetries = 17;
     int maxSubsequentStateExecutions = 100;
     Map<WorkflowState, Integer> maxSubsequentStateExecutionsPerState = new HashMap<>();
-    ReadablePeriod historyDeletableAfter;
+    ReadablePeriod historyDeletableAfter = Period.days(45);
     short defaultPriority = 0;
     BooleanSupplier deleteHistoryCondition = onAverageEveryNthExecution(100);
     BiFunction<WorkflowState, Throwable, StateProcessExceptionHandling> exceptionAnalyzer;
@@ -115,7 +113,9 @@ public class WorkflowSettings extends ModelObject {
     /**
      * Returns true randomly every n:th time.
      *
-     * @param n The frequency of returning true.
+     * @param n
+     *          Controls the frequency of returning true. With n=1, returns true every time. With n=100, returns true on average
+     *          once per 100 calls.
      * @return Producer of boolean values
      */
     @SuppressFBWarnings(value = "MDM_RANDOM_SEED", justification = "Random does not need to be secure here")
@@ -124,18 +124,18 @@ public class WorkflowSettings extends ModelObject {
     }
 
     /**
-     * Returns true randomly once per day (during the early hours).
+     * Returns true once per day (during early hours).
      *
      * @return Producer of boolean values
      */
     public static BooleanSupplier oncePerDay() {
-      // this minutes and seconds vary by start time between nodes
+      // minutes and seconds vary by start time between nodes
       AtomicLong nextExecution = new AtomicLong(LocalDateTime.now().plusDays(1).withHourOfDay(4).toDateTime().getMillis());
       return () -> {
         long now = currentTimeMillis();
         long next = nextExecution.get();
         if (now > next) {
-          nextExecution.set(next + DAYS.toMillis(1));
+          nextExecution.set(next + standardDays(1).getMillis());
           return true;
         }
         return false;
@@ -146,10 +146,10 @@ public class WorkflowSettings extends ModelObject {
      * Set the maximum delay on execution retry after an error.
      *
      * @param maxErrorTransitionDelay
-     *          Delay in milliseconds.
+     *          The delay.
      * @return this.
      */
-    public Builder setMaxErrorTransitionDelay(int maxErrorTransitionDelay) {
+    public Builder setMaxErrorTransitionDelay(Duration maxErrorTransitionDelay) {
       this.maxErrorTransitionDelay = maxErrorTransitionDelay;
       return this;
     }
@@ -158,10 +158,10 @@ public class WorkflowSettings extends ModelObject {
      * Set the minimum delay on execution retry after an error.
      *
      * @param minErrorTransitionDelay
-     *          Delay in milliseconds.
+     *          The delay.
      * @return this.
      */
-    public Builder setMinErrorTransitionDelay(int minErrorTransitionDelay) {
+    public Builder setMinErrorTransitionDelay(Duration minErrorTransitionDelay) {
       this.minErrorTransitionDelay = minErrorTransitionDelay;
       return this;
     }
@@ -170,23 +170,11 @@ public class WorkflowSettings extends ModelObject {
      * Set the length of forced delay to break execution of a step that is considered to be busy looping.
      *
      * @param shortTransitionDelay
-     *          Delay in milliseconds.
+     *          The delay.
      * @return this.
      */
-    public Builder setShortTransitionDelay(int shortTransitionDelay) {
+    public Builder setShortTransitionDelay(Duration shortTransitionDelay) {
       this.shortTransitionDelay = shortTransitionDelay;
-      return this;
-    }
-
-    /**
-     * Set immediate transition delay.
-     *
-     * @param immediateTransitionDelay
-     *          Delay in milliseconds.
-     * @return this.
-     */
-    public Builder setImmediateTransitionDelay(int immediateTransitionDelay) {
-      this.immediateTransitionDelay = immediateTransitionDelay;
       return this;
     }
 
@@ -229,8 +217,8 @@ public class WorkflowSettings extends ModelObject {
     }
 
     /**
-     * Set the delay after which workflow history (actions, states) can be deleted from the database by nFlow.
-     * The default value (<code>null</code>) indicates that history is not deletable.
+     * Set the delay after which workflow history (actions, states) can be deleted from the database by nFlow. Setting value to
+     * <code>null</code> indicates that history is not deletable. Default is 45 days.
      *
      * @param period
      *          Delay after which history can be deleted.
@@ -311,12 +299,11 @@ public class WorkflowSettings extends ModelObject {
    */
   protected long calculateBinaryBackoffDelay(int retryCount, long minDelay, long maxDelay) {
     BigInteger delay = BigInteger.valueOf(minDelay).multiply(BigInteger.valueOf(2).pow(retryCount));
-    if(!BigInteger.valueOf(delay.longValue()).equals(delay)) {
-      // got overflow in delay calculation
-      // Java 1.8 has delay.longValueExact()
+    try {
+      return max(minDelay, min(delay.longValueExact(), maxDelay));
+    } catch (@SuppressWarnings("unused") ArithmeticException overflow) {
       return maxDelay;
     }
-    return max(minDelay, min(delay.longValue(), maxDelay));
   }
 
   /**
@@ -325,7 +312,7 @@ public class WorkflowSettings extends ModelObject {
    * @return The delay in milliseconds.
    */
   public DateTime getShortTransitionActivation() {
-    return now().plusMillis(shortTransitionDelay);
+    return now().plus(millis(shortTransitionDelay));
   }
 
   /**