+++ nflow-rest-api-common/src/main/java/io/nflow/rest/v1/converter/ListWorkflowDefinitionConverter.java
@@ -48,7 +48,6 @@ public class ListWorkflowDefinitionConverter {
 
     WorkflowSettings workflowSettings = definition.getSettings();
     TransitionDelays transitionDelays = new TransitionDelays();
-    transitionDelays.immediate = workflowSettings.immediateTransitionDelay;
     transitionDelays.waitShort = workflowSettings.shortTransitionDelay;
     transitionDelays.minErrorWait = workflowSettings.minErrorTransitionDelay;
     transitionDelays.maxErrorWait = workflowSettings.maxErrorTransitionDelay;