+++ nflow-rest-api-common/src/test/java/io/nflow/rest/v1/DummyTestWorkflow.java
@@ -4,6 +4,7 @@ import static io.nflow.engine.workflow.definition.NextAction.moveToState;
 import static io.nflow.rest.v1.TestState.BEGIN;
 import static io.nflow.rest.v1.TestState.DONE;
 import static io.nflow.rest.v1.TestState.ERROR;
+import static org.joda.time.Duration.millis;
 import static org.joda.time.Period.days;
 
 import java.util.HashMap;
@@ -17,9 +18,9 @@ import io.nflow.engine.workflow.definition.WorkflowSettings;
 public class DummyTestWorkflow extends AbstractWorkflowDefinition {
 
   public DummyTestWorkflow() {
-    super("dummy", BEGIN, ERROR, new WorkflowSettings.Builder().setMinErrorTransitionDelay(300).setMaxErrorTransitionDelay(1000)
-        .setShortTransitionDelay(200).setImmediateTransitionDelay(100).setMaxRetries(10).setHistoryDeletableAfter(days(3))
-        .build());
+    super("dummy", BEGIN, ERROR,
+        new WorkflowSettings.Builder().setMinErrorTransitionDelay(millis(300)).setMaxErrorTransitionDelay(millis(1000))
+            .setShortTransitionDelay(millis(200)).setMaxRetries(10).setHistoryDeletableAfter(days(3)).build());
     permit(BEGIN, DONE, ERROR);
     permit(BEGIN, ERROR);
     permit(ERROR, DONE);