+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 - `nflow-engine`
   - BREAKING CHANGE: Remove `WorkflowDefinition`, workflow definitions should extend `AbstractWorkflowDefinition` instead.
   - BREAKING CHANGE: Remove deprecated `WorkflowState.isRetryAllowed`, set exception analyzer for workflow definition instead (if needed).
+  - BREAKING CHANGE: Change transition delay fields data types in `WorkflowSettings` and remove unused `immediateTransitionDelay` setting.
+- `nflow-rest-api`
+  - BREAKING CHANGE: Remove `ListWorkflowDefinitionResponse.TransitionDelays.immediate` field, it is not used by nFlow anymore.
+  - Enable maintenance (archiving and deleting old workflow instances) by default.
+  - Enable workflow instance history clean-up (deleting old actions and state variables) by default.
+  - Add support to query also archived workflow instances.
 
 **Details**
 - `nflow-engine`
@@ -12,6 +18,13 @@
     - You can define the states as instances of `io.nflow.engine.workflow.curated.State` or anything else that implements the required `io.nflow.engine.workflow.definition.WorkflowState` interface.
     - The workflow definitions must now register all possible states as described in `io.nflow.engine.workflow.definition.AbstractWorkflowDefinition`.
   - `WorkflowState.isRetryAllowed` was removed. If it was overridden, you can use `new WorkflowSettings.Builder().setExceptionAnalyzer(...)` to change the behavior. The default behavior was not changed.
+  - `WorkflowSettings`
+    - Change `WorkflowSettings.Builder.setShortTransitionDelay`, `WorkflowSettings.Builder.setMinErrorTransitionDelay` and `WorkflowSettings.Builder.setMaxErrorTransitionDelay` parameter type from `int` to `org.joda.time.Duration`.
+    - Remove `WorkflowSettings.Builder.setImmediateTransitionDelay` method.
+    - Change `WorkflowSettings.shortTransitionDelay`, `WorkflowSettings.minErrorTransitionDelay` and `WorkflowSettings.maxErrorTransitionDelay` field type from `int` to `long`.
+    - Remove `WorkflowSettings.immediateTransitionDelay` field. It is not used by nFlow.
+  - Maintenance workflow instance is added to nFlow database by default in startup. Instances that have been in final state longer than 45 days are archived. Archived instances that have been in final state longer than one year are deleted. Maintenance is run every night. Use `nflow.maintenance` configuration options to change the defaults before startup, or update the maintenance workflow instance state variables after the instance has been created.
+  - Workflow instance actions and state variables that are older than 45 days are automatically cleaned up occasionally when the instance is processed. Use workflow settings to change the default time period (`setHistoryDeletableAfter`) and condition (`setDeleteHistoryCondition`) of the clean-up.
   - Add support to query also archived workflow instances when not enough non-archived matches are found.
   - Dependency updates
     - logback-classic update to version 1.2.10
@@ -26,9 +39,9 @@
     - javassist 3.28.0
     - jodatime 2.10.3
     - slf4j 1.7.33
-- `nflow-rest-api-jax-rs` and `nflow-rest-api-spring-web`
-    - Add `queryArchive=true` query parameter to query also archived workflow instances when not enough non-archived matches are found.
 - `nflow-rest-api`
+  - BREAKING CHANGE: Remove `ListWorkflowDefinitionResponse.TransitionDelays.immediate` field, it is not used by nFlow.
+  - Add `queryArchive=true` query parameter to query also archived workflow instances when not enough non-archived matches are found.
   - Dependency updates
     - swagger 1.6.4
 - `nflow-jetty`
@@ -42,6 +55,7 @@
   - Dependency updates
     - metrics 4.2.7
 - `nflow-explorer`
+  - Query and show also archived workflow instances when not enough non-archived matches are found. Querying and showing archived instances can be disabled in `config.js`.
   - Dependency updates
     - nodejs 16.13.2
     - npm 8.1.2
@@ -56,8 +70,6 @@
     - mysql 8.0.28
     - mariadb 2.7.5
     - postgresql 42.3.1
-- `nflow-explorer`
-  - Query and show also archived workflow instances when not enough non-archived matches are found. Querying and showing archived instances can be disabled in `config.js`.
 - Minimum supported Maven version for building is 3.6
 
 ## 7.4.0 (2021-12-27)